Done 1, 3, 4, 5 and 6 tasks
Task 1. On login screen.
Task 3. Transition from Login screen to Main screen, is displayed on
RetrievedData fragment.
Task 4. On Main screen.
Task 5. On Channels fragment.
Task 6. On Channels fragment.

To be honest, task 2 is done too, but it is not so good visible. To see
animations you should log in and click on first menu item without
picking on a list item. See appbar text animation.
When you pick a list item and channels list is visible, click on a
search icon on a appbar. See icon animation (is really short and hard to
see but exists)